requirements.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
pillow
numpy
tqdm
ftfy
